% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-poisson-pkgdown.R
\name{use_poisson_pkgdown}
\alias{use_poisson_pkgdown}
\title{Set up poissontemplate usage}
\usage{
use_poisson_pkgdown(pkg = getwd())
}
\arguments{
\item{pkg}{Path to package}
}
\value{
None
}
\description{
This function
\itemize{
\item registers poissontemplate usage in the pkgdown config;
\item explains what the next steps are.
}
}
